locking istio deps on its own golang package. part 1
